requirements.txt
appnope==0.1.2
attrs==21.2.0
backcall==0.2.0
cattrs==1.7.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
crowd-kit==0.0.4
decorator==5.0.9
idna==2.10
ipyplot==1.1.0
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
matplotlib-inline==0.1.2
nltk==3.6.2
numpy==1.21.0
pandas==1.3.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
regex==2021.7.6
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
shortuuid==1.0.1
simplejson==3.17.3
six==1.16.0
threadpoolctl==2.2.0
toloka-kit==0.1.9
tqdm==4.61.2
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
